+++ R/addins.R
@@ -9,9 +9,12 @@ addin_add_template <- function() {
     if (rstudioapi::hasFun("showPrompt")) {
       template <- rstudioapi::showPrompt(
         title = "Enter the flat template type",
-        message = "e.g. add = additional; min = minimal;",
+        message = "e.g. add = additional; dev = dev_history; teach = teaching",
         default = "additional"
       )
+      if (is.null(template)) {
+        return(NULL)
+      }
     } else {
       template <- readline("Enter the flat template type: ")
     }
@@ -21,6 +24,9 @@ addin_add_template <- function() {
         message = "Name is used to pre-fill the template",
         default = "my_fun"
       )
+      if (is.null(flat_name)) {
+        return(NULL)
+      }
     } else {
       flat_name <- readline("Enter the function name: ")
     }