+++ R/fill_description.R
@@ -25,11 +25,10 @@
 #'       "Everything can be set from a Rmarkdown file in your project."
 #'     ),
 #'     `Authors@R` = c(
-#'       person("Sebastien", "Rochette",
-#'         email = "sebastien@thinkr.fr",
-#'         role = c("aut", "cre"), comment = c(ORCID = "0000-0002-1565-9313")
-#'       ),
-#'       person(given = "ThinkR", role = "cph")
+#'       person("John", "Doe",
+#'         email = "john@email.me",
+#'         role = c("aut", "cre"), comment = c(ORCID = "0000-0000-0000-0000")
+#'       )
 #'     )
 #'   )
 #' )
@@ -41,6 +40,15 @@ fill_description <- function(pkg = ".", fields, overwrite = FALSE) {
   # on.exit(setwd(old))
 
   path <- normalizePath(pkg)
+
+  project_name <- get_pkg_name(pkg = pkg)
+  clean_pkg_name <- asciify_name(project_name, to_pkg = TRUE)
+  if (project_name != clean_pkg_name) {
+    warning(
+      "Your package was renamed: `", clean_pkg_name,
+      "` as a package name should only contain letters, numbers and dots."
+    )
+  }
 
   desc_file <- file.path(path, "DESCRIPTION")
 
@@ -60,7 +68,7 @@ fill_description <- function(pkg = ".", fields, overwrite = FALSE) {
   # usethis::use_description(fields = fields)
 
   fields_new <- usethis::use_description_defaults(
-    package = basename(path),
+    package = clean_pkg_name, # basename(path),
     roxygen = TRUE,
     fields = fields
   )